+++ src/components/SymbolFillColorSelect.vue
@@ -2,7 +2,9 @@
 import { type NullableSymbolItem } from "@/types/constants";
 import { computed } from "vue";
 import SymbolCodeSelect from "@/components/SymbolCodeSelect.vue";
-import { SYMBOL_FILL_COLORS } from "@/config/colors.ts";
+import { SYMBOL_FILL_COLORS, type SymbolFillColor } from "@/config/colors.ts";
+import { injectStrict } from "@/utils";
+import { activeScenarioKey } from "@/components/injects.ts";
 
 interface Props {
   sid?: string;
@@ -11,11 +13,16 @@ interface Props {
 
 const props = withDefaults(defineProps<Props>(), { sid: "3" });
 const emit = defineEmits(["update:modelValue"]);
-
 const colorValue = defineModel<string | null>({ default: null });
 
+const scn = injectStrict(activeScenarioKey);
+
 const colorIconItems = computed((): NullableSymbolItem[] =>
-  [{ code: null, text: "Default" }, ...SYMBOL_FILL_COLORS].map((item) => ({
+  [
+    { code: null, text: "Default" },
+    ...SYMBOL_FILL_COLORS,
+    ...Object.values(scn.store.state.symbolFillColorMap),
+  ].map((item) => ({
     ...item,
     sidc: "100" + props.sid + 10 + "00" + "00" + "0000000000",
     symbolOptions: item.code